+++ src/main/scala/rps/Main.scala
@@ -15,6 +15,35 @@ import wiro.Config
 import wiro.server.akkaHttp._
 import wiro.server.akkaHttp.FailSupport._
 
+import akka.actor.ActorSystem
+import akka.stream.ActorMaterializer
+import wiro.Config
+import wiro.server.akkaHttp._
+
+import rps.GameApi
+import akka.http.scaladsl.model._
+
+
+object WebServer extends App with RouterDerivationModule {
+  implicit val system = ActorSystem()
+  implicit val materializer = ActorMaterializer()
+  implicit val ec = system.dispatcher
+
+  implicit def throwableResponse: ToHttpResponse[Throwable] = error => 
+     HttpResponse(
+      status = StatusCodes.InternalServerError,
+      entity = HttpEntity(
+        ContentType(MediaTypes.`application/json`),
+        s"""{"error":${error.getMessage()}}"""))
+
+  val gameRouter = deriveRouter[GameApi](new GameApiImpl)
+
+  new HttpRPCServer(
+    config = Config("localhost", 8080),
+    routers = List(gameRouter)
+  )
+}
+
 @enum trait Move {
   object Rock
   object Paper
@@ -27,53 +56,4 @@ import wiro.server.akkaHttp.FailSupport._
   object Draw
 }
 
-import akka.actor.ActorSystem
-import akka.stream.ActorMaterializer
-import wiro.Config
-import wiro.server.akkaHttp._
-
-object WebServer extends App with RouterDerivationModule {
-  implicit val system = ActorSystem()
-  implicit val materializer = ActorMaterializer()
-  implicit val ec = system.dispatcher
-
-  implicit def throwableResponse: ToHttpResponse[Throwable] = null
-
-  val gameRouter = deriveRouter[GameApi](new GameApiImpl)
-
-  val rpcServer = new HttpRPCServer(
-    config = Config("localhost", 8080),
-    routers = List(gameRouter)
-  )
-}
-
-@path("rps")
-trait GameApi {
-  @command
-  def play(userMove:Move): Future[Either[Throwable, Response]]
-}
-
-class GameApiImpl(implicit ec: ExecutionContext) extends GameApi {
-
-  override def play(userMove: Move): Future[Either[Throwable, Response]] = Future {
-    val computerMove = generateRandom()
-    val result =(userMove, computerMove) match {
-        case (x,y) if x == y => Draw
-        case (Rock,Paper) | (Paper,Scissors) | (Scissors,Rock) => Lose
-        case (Rock,Scissors) | (Paper,Rock) | (Scissors,Paper) => Win   
-    }    
-    Right(Response tupled (userMove,computerMove,result))
-  }
-
-  def generateRandom() = {
-    Random.nextInt(3) match {
-      case 0 => Rock
-      case 1 => Paper
-      case 2 => Scissors
-      case _ => throw new Exception("errorissimo")
-    }
-  }
-}
-
-case class Request (userMove : Move)
 case class Response (userMove: Move, computerMove: Move, result: Result)
